offloaded card creation and updated styles
