+++ scripts/scripts.js
@@ -11,6 +11,7 @@ import {
   waitForLCP,
   loadBlocks,
   loadCSS,
+  createOptimizedPicture,
 } from './lib-franklin.js';
 
 const LCP_BLOCKS = ['header', 'hero']; // add your LCP blocks to the list
@@ -122,6 +123,50 @@ export async function lookupBlogs(pathnames) {
   }
   const result = pathnames.map((path) => window.blogIndex.lookup[path]).filter((e) => e);
   return (result);
+}
+
+/**
+ * Creates a Card using a JSON object and style associated with the card
+ * @param {Object} row JSON Object typically coming from an index array item
+ * @param {String} style Class name that needs to be added to the card root div
+ */
+
+export async function createCard(row, style) {
+  // Create card div
+  const card = document.createElement('div');
+  if (style) card.classList.add(style);
+
+  // Add the image to the card first
+  if (row.image !== '0' && row.title !== '0') {
+    const cardImage = document.createElement('div');
+    cardImage.classList.add('card-image');
+    cardImage.append(createOptimizedPicture(row.image, row.title));
+    card.prepend(cardImage);
+  }
+
+  // Create a separate child div for the card content
+  const cardContent = document.createElement('div');
+  cardContent.classList.add('card-content');
+
+  // Create and add the link, title, author, readtime and category to card content and card
+  const link = document.createElement('a');
+  link.classList.add('blog-link');
+  link.href = row.path;
+  if (row.title) link.innerHTML += `<h5>${row.title}</h5>`;
+  cardContent.append(link);
+  if (row.description && row.description !== '0') cardContent.innerHTML += `<p>${row.description}</p>`;
+  const author = document.createElement('div');
+  author.classList.add('blog-author');
+  if (row.author && row.author !== '0') author.innerHTML += `By ${row.author}`;
+  if (row.readtime && row.readtime !== '0') author.innerHTML += ` | ${row.readtime}`;
+  cardContent.append(author);
+  const category = document.createElement('div');
+  category.classList.add('blog-category');
+  if (row.category && row.category !== '0') category.innerHTML += row.category;
+  cardContent.append(category);
+
+  card.append(cardContent);
+  return (card);
 }
 
 /**
