+++ styles/styles.css
@@ -95,6 +95,7 @@
   --spacer-layout-01: 16px;
   --spacer-layout-02: 24px;
   --spacer-layout-03: 32px;
+  --spacer-layout-036: 36px;
   --spacer-layout-04: 48px;
   --spacer-layout-05: 64px;
   --spacer-layout-06: 96px;
